requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.3
oauthlib==2.0.7
pandas==0.22.0
pkg-resources==0.0.0
psycopg2==2.7.4
python-dateutil==2.7.2
python-twitter==3.4.1
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
Werkzeug==0.14.1
